force int for conversion

+++ src/inc/utils/AgentUtils.class.php
@@ -43,7 +43,7 @@ class AgentUtils {
       }
       $data = explode(",", $entry->getValue());
       for($i = 0; $i < sizeof($data); $i++){
-        $pos = $i + sizeof($data)*array_search($entry->getStatType(), $types);
+        $pos = (int)($i + sizeof($data)*array_search($entry->getStatType(), $types));
         if(!isset($datasets[$pos])){
           $datasets[$pos] = array(
             "label" => "Device #" . ($i + 1),
